remove redundant cast
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1555719 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/facet/src/test/org/apache/lucene/facet/range/TestRangeFacetCounts.java
@@ -541,8 +541,8 @@ public class TestRangeFacetCounts extends FacetTestCase {
         // Must truncate to float precision so that the
         // drill-down counts (which use NRQ.newFloatRange)
         // are correct:
-        min = (double) (float) min;
-        max = (double) (float) max;
+        min = (float) min;
+        max = (float) max;
 
         boolean minIncl;
         boolean maxIncl;
